Following changes were done 1) Made auth & schema module independent of each other. 2) Made crud module independent of Auth & Schema.
