+++ gateway/modules/auth/helpers/helpers.go
@@ -1,0 +1,17 @@
+package helpers
+
+import (
+	"crypto/aes"
+	"crypto/cipher"
+)
+
+// DecryptAESCFB decrypts aes cfb string
+func DecryptAESCFB(dst, src, key, iv []byte) error {
+	aesBlockDecrypter, err := aes.NewCipher([]byte(key))
+	if err != nil {
+		return err
+	}
+	aesDecrypter := cipher.NewCFBDecrypter(aesBlockDecrypter, iv)
+	aesDecrypter.XORKeyStream(dst, src)
+	return nil
+}
