+++ gateway/modules/crud/operations.go
@@ -7,6 +7,7 @@ import (
 	"github.com/spaceuptech/helpers"
 
 	"github.com/spaceuptech/space-cloud/gateway/model"
+	schemaHelpers "github.com/spaceuptech/space-cloud/gateway/modules/schema/helpers"
 	"github.com/spaceuptech/space-cloud/gateway/utils"
 )
 
@@ -15,12 +16,12 @@ func (m *Module) Create(ctx context.Context, dbAlias, col string, req *model.Cre
 	m.RLock()
 	defer m.RUnlock()
 
-	if err := m.schema.ValidateCreateOperation(ctx, dbAlias, col, req); err != nil {
+	crud, err := m.getCrudBlock(dbAlias)
+	if err != nil {
 		return err
 	}
 
-	crud, err := m.getCrudBlock(dbAlias)
-	if err != nil {
+	if err := schemaHelpers.ValidateCreateOperation(ctx, dbAlias, string(crud.GetDBType()), col, m.schemaDoc, req); err != nil {
 		return err
 	}
 
@@ -60,7 +61,7 @@ func (m *Module) Read(ctx context.Context, dbAlias, col string, req *model.ReadR
 	}
 
 	// Adjust where clause
-	if err := m.schema.AdjustWhereClause(ctx, dbAlias, crud.GetDBType(), col, req.Find); err != nil {
+	if err := schemaHelpers.AdjustWhereClause(ctx, dbAlias, crud.GetDBType(), col, m.schemaDoc, req.Find); err != nil {
 		return nil, nil, err
 	}
 
@@ -89,7 +90,7 @@ func (m *Module) Read(ctx context.Context, dbAlias, col string, req *model.ReadR
 	n, result, _, metaData, err := crud.Read(ctx, col, req)
 
 	// Process the response
-	if err := m.schema.CrudPostProcess(ctx, dbAlias, col, result); err != nil {
+	if err := schemaHelpers.CrudPostProcess(ctx, dbAlias, string(crud.GetDBType()), col, m.schemaDoc, result); err != nil {
 		return nil, nil, helpers.Logger.LogError(helpers.GetRequestID(ctx), fmt.Sprintf("error executing read request in crud module unable to perform schema post process for un marshalling json for project (%s) col (%s)", m.project, col), err, nil)
 	}
 
@@ -110,12 +111,12 @@ func (m *Module) Update(ctx context.Context, dbAlias, col string, req *model.Upd
 	m.RLock()
 	defer m.RUnlock()
 
-	if err := m.schema.ValidateUpdateOperation(ctx, dbAlias, col, req.Operation, req.Update, req.Find); err != nil {
+	crud, err := m.getCrudBlock(dbAlias)
+	if err != nil {
 		return err
 	}
 
-	crud, err := m.getCrudBlock(dbAlias)
-	if err != nil {
+	if err := schemaHelpers.ValidateUpdateOperation(ctx, dbAlias, string(crud.GetDBType()), col, req.Operation, req.Update, req.Find, m.schemaDoc); err != nil {
 		return err
 	}
 
@@ -124,7 +125,7 @@ func (m *Module) Update(ctx context.Context, dbAlias, col string, req *model.Upd
 	}
 
 	// Adjust where clause
-	if err := m.schema.AdjustWhereClause(ctx, dbAlias, crud.GetDBType(), col, req.Find); err != nil {
+	if err := schemaHelpers.AdjustWhereClause(ctx, dbAlias, crud.GetDBType(), col, m.schemaDoc, req.Find); err != nil {
 		return err
 	}
 
@@ -154,7 +155,7 @@ func (m *Module) Delete(ctx context.Context, dbAlias, col string, req *model.Del
 	}
 
 	// Adjust where clause
-	if err := m.schema.AdjustWhereClause(ctx, dbAlias, crud.GetDBType(), col, req.Find); err != nil {
+	if err := schemaHelpers.AdjustWhereClause(ctx, dbAlias, crud.GetDBType(), col, m.schemaDoc, req.Find); err != nil {
 		return err
 	}
 
@@ -230,25 +231,25 @@ func (m *Module) Batch(ctx context.Context, dbAlias string, req *model.BatchRequ
 	m.RLock()
 	defer m.RUnlock()
 
+	crud, err := m.getCrudBlock(dbAlias)
+	if err != nil {
+		return err
+	}
+
 	for _, r := range req.Requests {
 		switch r.Type {
 		case string(model.Create):
 			v := &model.CreateRequest{Document: r.Document, Operation: r.Operation}
-			if err := m.schema.ValidateCreateOperation(ctx, dbAlias, r.Col, v); err != nil {
+			if err := schemaHelpers.ValidateCreateOperation(ctx, dbAlias, string(crud.GetDBType()), r.Col, m.schemaDoc, v); err != nil {
 				return err
 			}
 			r.Document = v.Document
 			r.Operation = v.Operation
 		case string(model.Update):
-			if err := m.schema.ValidateUpdateOperation(ctx, dbAlias, r.Col, r.Operation, r.Update, r.Find); err != nil {
+			if err := schemaHelpers.ValidateUpdateOperation(ctx, dbAlias, string(crud.GetDBType()), r.Col, r.Operation, r.Update, r.Find, m.schemaDoc); err != nil {
 				return err
 			}
 		}
-	}
-
-	crud, err := m.getCrudBlock(dbAlias)
-	if err != nil {
-		return err
 	}
 
 	if err := crud.IsClientSafe(ctx); err != nil {
@@ -359,4 +360,27 @@ func (m *Module) IsPreparedQueryPresent(dbAlias, id string) bool {
 	defer m.RUnlock()
 	_, p := m.queries[getPreparedQueryKey(dbAlias, id)]
 	return p
+}
+
+// GetSchema function gets schema
+func (m *Module) GetSchema(dbAlias, col string) (model.Fields, bool) {
+	m.RLock()
+	defer m.RUnlock()
+
+	dbSchema, p := m.schemaDoc[dbAlias]
+	if !p {
+		return nil, false
+	}
+
+	colSchema, p := dbSchema[col]
+	if !p {
+		return nil, false
+	}
+
+	fields := make(model.Fields, len(colSchema))
+	for k, v := range colSchema {
+		fields[k] = v
+	}
+
+	return fields, true
 }
