+++ gateway/modules/eventing/eventing_test.go
@@ -10,23 +10,16 @@ func TestModule_SetConfig(t *testing.T) {
 	type args struct {
 		eventing *config.EventingConfig
 	}
-	type mockArgs struct {
-		method         string
-		args           []interface{}
-		paramsReturned []interface{}
-	}
 	tests := []struct {
-		name           string
-		m              *Module
-		args           args
-		schemaMockArgs []mockArgs
-		wantErr        bool
+		name    string
+		m       *Module
+		args    args
+		wantErr bool
 	}{
 		{
-			name:           "eventing is not enabled",
-			m:              &Module{config: &config.Eventing{Enabled: true}},
-			args:           args{eventing: &config.EventingConfig{Enabled: false, DBAlias: "mysql"}},
-			schemaMockArgs: []mockArgs{},
+			name: "eventing is not enabled",
+			m:    &Module{config: &config.Eventing{Enabled: true}},
+			args: args{eventing: &config.EventingConfig{Enabled: false, DBAlias: "mysql"}},
 		},
 		{
 			name:    "DBAlias not mentioned",
@@ -37,20 +30,9 @@ func TestModule_SetConfig(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-
-			mockSchema := &mockSchemaEventingInterface{}
-
-			for _, m := range tt.schemaMockArgs {
-				mockSchema.On(m.method, m.args...).Return(m.paramsReturned...)
-			}
-
-			tt.m.schema = mockSchema
-
 			if err := tt.m.SetConfig("projectID", tt.args.eventing); (err != nil) != tt.wantErr {
 				t.Errorf("Module.SetConfig() error = %v, wantErr %v", err, tt.wantErr)
 			}
-
-			mockSchema.AssertExpectations(t)
 		})
 	}
 }
