+++ gateway/modules/eventing/types.go
@@ -33,8 +33,14 @@ type mockCrudInterface struct {
 	mock.Mock
 }
 
+func (m *mockCrudInterface) GetSchema(dbAlias, col string) (model.Fields, bool) {
+	c := m.Called(dbAlias, col)
+	return c.Get(0).(model.Fields), c.Bool(1)
+}
+
 func (m *mockCrudInterface) GetDBType(dbAlias string) (string, error) {
-	panic("implement me")
+	c := m.Called(dbAlias)
+	return c.String(0), c.Error(1)
 }
 
 func (m *mockCrudInterface) InternalCreate(ctx context.Context, dbAlias, project, col string, req *model.CreateRequest, isIgnoreMetrics bool) error {
@@ -121,44 +127,6 @@ func (m *mockAuthEventingInterface) GetSCAccessToken(context.Context) (string, e
 func (m *mockAuthEventingInterface) GetInternalAccessToken(context.Context) (string, error) {
 	c := m.Called()
 	return c.String(0), c.Error(1)
-}
-
-type mockSchemaEventingInterface struct {
-	mock.Mock
-}
-
-func (m *mockSchemaEventingInterface) GetSchemaForDB(ctx context.Context, dbAlias, col, format string) ([]interface{}, error) {
-	c := m.Called(ctx, dbAlias, col, format)
-	return c.Get(0).([]interface{}), c.Error(1)
-}
-
-func (m *mockSchemaEventingInterface) SchemaInspection(ctx context.Context, dbAlias, project, col string) (string, error) {
-	c := m.Called(ctx, dbAlias, project, col)
-	return c.String(0), c.Error(1)
-}
-
-func (m *mockSchemaEventingInterface) CheckIfEventingIsPossible(dbAlias, col string, obj map[string]interface{}, isFind bool) (findForUpdate map[string]interface{}, present bool) {
-	c := m.Called(dbAlias, col, obj, isFind)
-	return map[string]interface{}{}, c.Bool(1)
-}
-
-func (m *mockSchemaEventingInterface) Parser(dbSchemas config.DatabaseSchemas) (model.Type, error) {
-	c := m.Called(dbSchemas)
-	return nil, c.Error(1)
-}
-
-func (m *mockSchemaEventingInterface) SchemaValidator(ctx context.Context, dbAlias, col string, collectionFields model.Fields, doc map[string]interface{}) (map[string]interface{}, error) {
-	c := m.Called(ctx, dbAlias, col, collectionFields, doc)
-	return nil, c.Error(1)
-}
-
-func (m *mockSchemaEventingInterface) SchemaModifyAll(ctx context.Context, dbAlias, logicalDBName string, tables config.DatabaseSchemas) error {
-	c := m.Called(ctx, dbAlias, logicalDBName, tables)
-	return c.Error(0)
-}
-func (m *mockSchemaEventingInterface) GetSchema(dbAlias, col string) (model.Fields, bool) {
-	c := m.Called(dbAlias, col)
-	return c.Get(0).(model.Fields), c.Bool(1)
 }
 
 type mockFileStoreEventingInterface struct {
