+++ gateway/modules/schema/creation_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/spaceuptech/space-cloud/gateway/config"
 	"github.com/spaceuptech/space-cloud/gateway/model"
 	"github.com/spaceuptech/space-cloud/gateway/modules/crud"
+	. "github.com/spaceuptech/space-cloud/gateway/modules/schema/helpers"
 )
 
 func TestSchema_generateCreationQueries(t *testing.T) {
@@ -295,8 +296,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "NO ACTION"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "NO ACTION"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "NO ACTION"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
+				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "NO ACTION"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
 			want:    []string{},
@@ -308,8 +309,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "CASCADE"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "CASCADE"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "CASCADE"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
+				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "CASCADE"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
 			want:    []string{},
@@ -491,7 +492,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
@@ -504,7 +505,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col2"), To: "id"}}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col2"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
@@ -517,7 +518,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table1", "col2"), OnDelete: "CASCADE"}}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table1", "col2"), OnDelete: "CASCADE"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
@@ -530,8 +531,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}}},
-				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}}},
+				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
 			want:    []string{"ALTER TABLE table1 DROP FOREIGN KEY c_table1_col1", "ALTER TABLE table1 DROP INDEX c_table1_col1", "ALTER TABLE table1 ADD CONSTRAINT c_table1_col1 FOREIGN KEY (col1) REFERENCES table2 (id) ON DELETE CASCADE"},
@@ -543,8 +544,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
-				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
+				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, Args: &model.FieldArgs{Precision: model.DefaultPrecision, Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
 			want:    []string{"ALTER TABLE table1 DROP FOREIGN KEY c_table1_col1", "ALTER TABLE table1 DROP INDEX c_table1_col1", "ALTER TABLE table1 ADD CONSTRAINT c_table1_col1 FOREIGN KEY (col1) REFERENCES table2 (id)"},
@@ -556,7 +557,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeBoolean, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeBoolean, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
@@ -842,7 +843,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: false}}, "table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
@@ -855,7 +856,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "mysql",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: false}}, "table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
@@ -869,7 +870,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				tableName:     "table1",
 				project:       "test",
 				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsFieldTypeRequired: false, IsForeign: false}}, "table2": model.Fields{"id": &model.FieldType{FieldName: "col1", TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
 			want:    []string{"ALTER TABLE table1 DROP FOREIGN KEY c_table1_col1", "ALTER TABLE table1 DROP INDEX c_table1_col1"},
@@ -882,7 +883,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				tableName:     "table1",
 				project:       "test",
 				parsedSchema:  model.Type{"mysql": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsFieldTypeRequired: false, IsForeign: false}}, "table2": model.Fields{"id": &model.FieldType{FieldName: "col1", TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudMySQL, project: "test"},
 			want:    []string{"ALTER TABLE table1 DROP FOREIGN KEY c_table1_col1", "ALTER TABLE table1 DROP INDEX c_table1_col1", "ALTER TABLE table1 DROP COLUMN col1", "ALTER TABLE table1 ADD col1 bigint"},
@@ -1447,8 +1448,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "NO ACTION"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "NO ACTION"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "NO ACTION"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
+				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "NO ACTION"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
 			want:    []string{},
@@ -1460,8 +1461,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "CASCADE"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "CASCADE"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "CASCADE"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
+				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "CASCADE"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
 			want:    []string{},
@@ -1617,7 +1618,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
@@ -1630,7 +1631,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col2"), To: "id"}}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col2"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
@@ -1643,7 +1644,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col2"), To: "id", OnDelete: "CASCADE"}}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col2"), To: "id", OnDelete: "CASCADE"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
@@ -1656,8 +1657,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}}},
-				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}}},
+				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1", "ALTER TABLE test.table1 ADD CONSTRAINT c_table1_col1 FOREIGN KEY (col1) REFERENCES test.table2 (id) ON DELETE CASCADE"},
@@ -1669,8 +1670,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
-				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
+				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1", "ALTER TABLE test.table1 ADD CONSTRAINT c_table1_col1 FOREIGN KEY (col1) REFERENCES test.table2 (id)"},
@@ -1682,7 +1683,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeBoolean, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeBoolean, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
@@ -1877,7 +1878,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: false}}, "table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
@@ -1890,7 +1891,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "sqlserver",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: false}}, "table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
@@ -1904,7 +1905,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				tableName:     "table1",
 				project:       "test",
 				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsFieldTypeRequired: false, IsForeign: false}}, "table2": model.Fields{"id": &model.FieldType{FieldName: "col1", TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1"},
@@ -1917,7 +1918,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				tableName:     "table1",
 				project:       "test",
 				parsedSchema:  model.Type{"sqlserver": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsFieldTypeRequired: false, IsForeign: false}}, "table2": model.Fields{"id": &model.FieldType{FieldName: "col1", TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudSQLServer, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1", "ALTER TABLE test.table1 DROP COLUMN col1", "ALTER TABLE test.table1 ADD col1 bigint"},
@@ -2601,8 +2602,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "NO ACTION"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "NO ACTION"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "NO ACTION"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
+				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "NO ACTION"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
 			want:    []string{},
@@ -2614,8 +2615,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "CASCADE"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: getConstraintName("table2", "id"), OnDelete: "CASCADE"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "CASCADE"}, TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
+				currentSchema: model.Collection{"table2": model.Fields{"id": &model.FieldType{FieldName: "id", IsFieldTypeRequired: true, IsPrimary: true}}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", IsForeign: true, JointTable: &model.TableProperties{Table: "table2", To: "id", ConstraintName: GetConstraintName("table2", "id"), OnDelete: "CASCADE"}, Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize}}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
 			want:    []string{},
@@ -2693,7 +2694,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
@@ -2706,7 +2707,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col2"), To: "id"}}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col2"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
@@ -2719,7 +2720,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col2"), To: "id", OnDelete: "CASCADE"}}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col2": &model.FieldType{FieldName: "col2", Kind: model.TypeDateTime, Args: &model.FieldArgs{Scale: model.DefaultScale}, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col2"), To: "id", OnDelete: "CASCADE"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
@@ -2732,8 +2733,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}}},
-				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}}},
+				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1", "ALTER TABLE test.table1 ADD CONSTRAINT c_table1_col1 FOREIGN KEY (col1) REFERENCES test.table2 (id) ON DELETE CASCADE"},
@@ -2745,8 +2746,8 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
-				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
+				currentSchema: model.Collection{"table2": model.Fields{}, "table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeDateTime, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id", OnDelete: "CASCADE"}}}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1", "ALTER TABLE test.table1 ADD CONSTRAINT c_table1_col1 FOREIGN KEY (col1) REFERENCES test.table2 (id)"},
@@ -2758,7 +2759,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeBoolean, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"id": &model.FieldType{FieldName: "id", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsPrimary: true, PrimaryKeyInfo: &model.TableProperties{}, IsFieldTypeRequired: true}, "col1": &model.FieldType{FieldName: "col1", Kind: model.TypeBoolean, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
@@ -3032,7 +3033,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				dbAlias:       "postgres",
 				tableName:     "table1",
 				project:       "test",
-				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}}},
+				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}}},
 				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: false}}, "table2": model.Fields{}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
@@ -3046,7 +3047,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				tableName:     "table1",
 				project:       "test",
 				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsFieldTypeRequired: false, IsForeign: false}}, "table2": model.Fields{"id": &model.FieldType{FieldName: "col1", TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1"},
@@ -3059,7 +3060,7 @@ func TestSchema_generateCreationQueries(t *testing.T) {
 				tableName:     "table1",
 				project:       "test",
 				parsedSchema:  model.Type{"postgres": model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeInteger, IsFieldTypeRequired: false, IsForeign: false}}, "table2": model.Fields{"id": &model.FieldType{FieldName: "col1", TypeIDSize: model.SQLTypeIDSize, Kind: model.TypeID}}}},
-				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: getConstraintName("table1", "col1"), To: "id"}}}},
+				currentSchema: model.Collection{"table1": model.Fields{"col1": &model.FieldType{FieldName: "col1", Kind: model.TypeID, TypeIDSize: model.SQLTypeIDSize, IsForeign: true, JointTable: &model.TableProperties{Table: "table2", ConstraintName: GetConstraintName("table1", "col1"), To: "id"}}}},
 			},
 			fields:  fields{crud: crudPostgres, project: "test"},
 			want:    []string{"ALTER TABLE test.table1 DROP CONSTRAINT c_table1_col1", "ALTER TABLE test.table1 DROP COLUMN col1", "ALTER TABLE test.table1 ADD COLUMN col1 bigint"},
