+++ gateway/modules/schema/operations_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/spaceuptech/space-cloud/gateway/config"
 	"github.com/spaceuptech/space-cloud/gateway/model"
+	. "github.com/spaceuptech/space-cloud/gateway/modules/schema/helpers"
 )
 
 func TestManager_GetSchemas(t *testing.T) {
@@ -106,7 +107,7 @@ func TestManager_GetSchemas(t *testing.T) {
 								Table:          "genres",
 								To:             "id",
 								OnDelete:       "NO ACTION",
-								ConstraintName: getConstraintName("authors", "genre_id"),
+								ConstraintName: GetConstraintName("authors", "genre_id"),
 							},
 						},
 					}},
@@ -178,7 +179,7 @@ func TestManager_GetSchemas(t *testing.T) {
 								Table:          "genres",
 								To:             "id",
 								OnDelete:       "NO ACTION",
-								ConstraintName: getConstraintName("authors", "genre_id"),
+								ConstraintName: GetConstraintName("authors", "genre_id"),
 							},
 						},
 					},
@@ -208,7 +209,7 @@ func TestManager_GetSchemas(t *testing.T) {
 								Table:          "authors",
 								To:             "id",
 								OnDelete:       "NO ACTION",
-								ConstraintName: getConstraintName("subscribers", "author_id"),
+								ConstraintName: GetConstraintName("subscribers", "author_id"),
 							},
 						},
 					},
@@ -272,7 +273,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "genres",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("authors", "genre_id"),
+		// 						ConstraintName: GetConstraintName("authors", "genre_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -304,7 +305,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "authors",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("subscribers", "author_id"),
+		// 						ConstraintName: GetConstraintName("subscribers", "author_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -353,7 +354,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "genres",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("authors", "genre_id"),
+		// 						ConstraintName: GetConstraintName("authors", "genre_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -385,7 +386,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "authors",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("subscribers", "author_id"),
+		// 						ConstraintName: GetConstraintName("subscribers", "author_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -436,7 +437,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "genres",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("authors", "genre_id"),
+		// 						ConstraintName: GetConstraintName("authors", "genre_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -468,7 +469,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "authors",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("subscribers", "author_id"),
+		// 						ConstraintName: GetConstraintName("subscribers", "author_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -517,7 +518,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "genres",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("authors", "genre_id"),
+		// 						ConstraintName: GetConstraintName("authors", "genre_id"),
 		// 					},
 		// 				},
 		// 			},
@@ -549,7 +550,7 @@ func TestManager_GetSchemas(t *testing.T) {
 		// 						Table:          "authors",
 		// 						To:             "id",
 		// 						OnDelete:       "NO ACTION",
-		// 						ConstraintName: getConstraintName("subscribers", "author_id"),
+		// 						ConstraintName: GetConstraintName("subscribers", "author_id"),
 		// 					},
 		// 				},
 		// 			},
